requirements.txt
aiogram>=3.4.0
SQLAlchemy>=2.0.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0